feat: prepared query for file upload && start of notifications

+++ backend/src/files/files.service.ts
@@ -6,9 +6,7 @@ import * as fs from "fs";
 import { AvatarFile } from "./models/avatar.model";
 import { DatabaseTable } from "../database/database.decorator";
 import { DatabaseService } from "../database/database.service";
-import { Event } from "../events/models/event.model";
 import { ConfigService } from "../config/config.service";
-import { f } from "@deepkit/type";
 
 @Injectable()
 export class FilesService {
@@ -21,6 +19,8 @@ export class FilesService {
       useProcess: true,
     });
   }
+
+  securityLevel: number = this.configService.databaseSecurity.securityLevel;
 
   create(createFileDto: CreateFileDto) {
     return "This action adds a new file";
@@ -35,10 +35,19 @@ export class FilesService {
   }
 
   async findOneAvatarFile(id: number): Promise<AvatarFile> {
-    return this.db.query({
-      query: "*",
-      where: "id = " + id,
-    });
+    if (this.securityLevel === 1 || this.securityLevel === 2) {
+      return this.db.query({
+        query: "*",
+        where: "id = " + id,
+      });
+    } else {
+      const where = "id = $1";
+      return this.db.preparedQuery({
+        query: "*",
+        where,
+        variables: [id],
+      });
+    }
   }
 
   update(id: number, updateFileDto: UpdateFileDto) {
@@ -49,33 +58,58 @@ export class FilesService {
     const { originalname, filename, mimetype, size } = file;
     const extension = mimetype.split("/")[1];
     const now = Date.now();
-    return await this.db.insert({
-      query:
-        "original_filename, current_filename, extension, size, created_at, updated_at",
-      where: `'${originalname}', '${filename}', '${extension}', ${size} ,${now}, ${now}`,
-    });
+    if (this.securityLevel === 1 || this.securityLevel === 2) {
+      return await this.db.insert({
+        query:
+          "original_filename, current_filename, extension, size, created_at, updated_at",
+        where: `'${originalname}', '${filename}', '${extension}', ${size} ,${now}, ${now}`,
+      });
+    } else {
+      const where = `$${1}, $${2}, $${3}, $${4}, $${5}, $${6}`;
+      return this.db.preparedInsert({
+        query:
+          "original_filename, current_filename, extension, size, created_at, updated_at",
+        where,
+        variables: [originalname, filename, extension, size, now, now],
+      });
+    }
   }
 
   async removeAvatarFile(pseudo: string, id: number): Promise<AvatarFile> {
     await this.unlinkAvatarFile(pseudo, id);
-    return this.db.delete({
-      query: "",
-      where: "id = " + id,
-    });
+    if (this.securityLevel === 1 || this.securityLevel === 2) {
+      return this.db.delete({
+        query: "",
+        where: "id = " + id,
+      });
+    } else {
+      const where = "id = $1";
+      return this.db.preparedDelete({
+        query: "",
+        where,
+        variables: [id],
+      });
+    }
   }
 
   async unlinkAvatarFile(pseudo: string, fileId: number) {
     const unlinkAsync = promisify(fs.unlink);
-    const file = await this.db.query({
-      query: "*",
-      where: "id = " + fileId,
-    });
+    let file = null;
+    if (this.securityLevel === 1 || this.securityLevel === 2) {
+      file = await this.db.query({
+        query: "*",
+        where: "id = " + fileId,
+      });
+    } else {
+      const where = "id = $1";
+      file = await this.db.preparedQuery({
+        query: "*",
+        where,
+        variables: [fileId],
+      });
+    }
+
     const { current_filename } = file;
-    console.log(
-      this.configService.getUploadPath() +
-        `/profil/${pseudo}/` +
-        current_filename
-    );
     await unlinkAsync(
       this.configService.getUploadPath() +
         `/profil/${pseudo}/` +
